requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.16
langchain-chroma>=0.0.1
pypdf>=3.17.1
chromadb>=0.4.22
tiktoken>=0.5.2
python-dotenv>=1.0.0
pydantic>=2.5.0 
